+++ README.md
@@ -4,11 +4,11 @@ This repository contains various components for interacting with an external app
 
 ## Environment
 
-`SocketAppEnv` in `components/env.py` provides an OpenAI Gym environment. By default it communicates with two separate UDP ports: one for sending actions and one for requesting rewards. When the environment is created with `combined_server=True`, both actions and reward requests are sent to the same address. This mode is intended for use with the `start_combined_udp_server` helper defined in `components/action_module.py`.
+`SocketAppEnv` in `envs/socket_env.py` provides an OpenAI Gym environment. By default it communicates with two separate UDP ports: one for sending actions and one for requesting rewards. When the environment is created with `combined_server=True`, both actions and reward requests are sent to the same address. This mode is intended for use with the `start_combined_udp_server` helper defined in `servers/action_server.py`.
 
 # Undertale RL via UDP
 
-This repository implements a reinforcement learning setup that interacts with an external application (e.g. Undertale) over UDP sockets. The main environment is `SocketAppEnv` in `components/env.py` which sends discrete actions and requests rewards from separate UDP endpoints. Observations are video frames encoded through a DINO model and combined with an intrinsic reward from `E3BIntrinsicReward`.
+This repository implements a reinforcement learning setup that interacts with an external application (e.g. Undertale) over UDP sockets. The main environment is `SocketAppEnv` in `envs/socket_env.py` which sends discrete actions and requests rewards from separate UDP endpoints. Observations are video frames encoded through a DINO model and combined with an intrinsic reward from `E3BIntrinsicReward`.
 
 ## Environment Workflow
 * Actions are sent over UDP to a keyboard server.
@@ -17,12 +17,12 @@ This repository implements a reinforcement learning setup that interacts with an
 * Extrinsic and intrinsic rewards are summed for each step.
 
 ## Action and Reward Server
-`components/action_module.py` exposes `start_combined_udp_server` which wraps `ExternalRewardTracker` from `external_reward.py` and processes both actions and reward queries on the same UDP port. Run this module to start the server.
+`servers/action_server.py` exposes `start_combined_udp_server` which wraps `ExternalRewardTracker` from `servers/reward_server.py` and processes both actions and reward queries on the same UDP port. Run this module to start the server.
 
 ## Neural Modules and PPO
-* `components/nn.py` contains LSTM based actor and critic networks.
-* `components/ppo.py` performs the PPO update using log‑prob ratios and clipping.
-* `components/rollout.py` provides a rollout buffer and GAE computation.
+* `models/nn.py` contains LSTM based actor and critic networks.
+* `models/ppo.py` performs the PPO update using log‑prob ratios and clipping.
+* `utils/rollout.py` provides a rollout buffer and GAE computation.
 
 ## Redundant Parts
 * `ExternalRewardTracker` exists in both `action_module.py` and `external_reward.py`.
@@ -35,8 +35,11 @@ This repository implements a reinforcement learning setup that interacts with an
 ## Running
 1. Start the combined UDP server:
    ```bash
-   python components/action_module.py
+   python servers/action_server.py
    ```
-2. Launch the training script (not provided yet – create one using the modules above).
+2. Launch the training script:
+   ```bash
+   python main.py
+   ```
 
 Dependencies include `torch`, `gymnasium`, `numpy`, `transformers`, `pymem`, `psutil`, `pynput` and `opencv-python`.
