+++ utils/intrinsic.py
