+++ utils/shrink.py
